Rename integration/split_data.py -> base.py
