+++ src/time_split/integration/base.py
